divided utility functions into customElements, dom, prefixes, serialization
