+++ package.json
@@ -2,7 +2,7 @@
   "name": "CETEIcean",
   "version": "1.9.5",
   "description": "JavaScript library to load a TEI XML document and register it as HTML5 custom elements.",
-  "main": "src/CETEI.js",
+  "main": "dist/CETEI.js",
   "type": "module",
   "keywords": [
     "TEI",
@@ -16,9 +16,16 @@
     "type": "git",
     "url": "https://github.com/TEIC/CETEIcean.git"
   },
+  "types": "dist/types/CETEI.d.ts",
   "exports": {
-    ".": "./src/CETEI.js",
-    "./utilities.js": "./src/utilities.js"
+    ".": {
+      "types": "./dist/types/CETEI.d.ts",
+      "default": "./dist/CETEI.js"
+    },
+    "./utilities.js": {
+      "types": "./dist/types/utilities.d.ts",
+      "default": "./dist/ts/utilities.js"
+    }
   },
   "devDependencies": {
     "@babel/core": "^7.26.9",
@@ -39,7 +46,7 @@
   },
   "scripts": {
     "build": "npm test && rollup -c rollup.config.js",
-    "build:tutorial": "npm run build && cp dist/CETEI.js tutorial/js/CETEI.js",
+    "build:tutorial": "npm run build && cp dist/CETEI.js tutorial_en/js/CETEI.js && cp dist/CETEI.js tutorial_es/js/CETEI.js && cp dist/CETEI.js tutorial_ja/js/CETEI.js",
     "dev": "npm run build && http-server -p 8888 & onchange src -- npm run build",
     "test": "tsx test/nodeTest.js",
     "typecheck": "tsc --noEmit"
